feat(#934): Redesign classes. TESTS FAILS
